test: move install PATH-guard test to its own clean file
ruff-changed lints the whole of any changed file, so appending to the
pre-existing-baseline-dirty test_installer.py surfaced its 6 grandfathered
findings and failed CI. Restore test_installer.py to its committed state and
put the PATH-guard assertions in a new, fully ruff-clean file instead.
